chore(cli): move getAutomaticNpmInstallSetting to commonUtils
