+++ packages/cli/tests/unit/cmds/preview/commonUtils.tests.ts
@@ -1,7 +1,13 @@
-import { IProgressHandler } from "@microsoft/teamsfx-api";
+import { IProgressHandler, err, ok, returnUserError } from "@microsoft/teamsfx-api";
 import sinon from "sinon";
-import { createTaskStartCb, createTaskStopCb } from "../../../../src/cmds/preview/commonUtils";
+import {
+  createTaskStartCb,
+  createTaskStopCb,
+  getAutomaticNpmInstallSetting,
+} from "../../../../src/cmds/preview/commonUtils";
 import { expect } from "../../utils";
+import { UserSettings } from "../../../../src/userSetttings";
+import { cliSource } from "../../../../src/constants";
 
 describe("commonUtils", () => {
   const sandbox = sinon.createSandbox();
@@ -29,6 +35,57 @@ describe("commonUtils", () => {
         exitCode: null,
       });
       expect(progressHandler.end.calledOnce).to.be.true;
+    });
+  });
+
+  describe("getAutomaticNpmInstallSetting", () => {
+    const automaticNpmInstallOption = "automatic-npm-install";
+
+    afterEach(() => {
+      sinon.restore();
+    });
+
+    it("on", () => {
+      sinon.stub(UserSettings, "getConfigSync").returns(
+        ok({
+          [automaticNpmInstallOption]: "on",
+        })
+      );
+      expect(getAutomaticNpmInstallSetting()).to.be.true;
+    });
+
+    it("off", () => {
+      sinon.stub(UserSettings, "getConfigSync").returns(
+        ok({
+          [automaticNpmInstallOption]: "off",
+        })
+      );
+      expect(getAutomaticNpmInstallSetting()).to.be.false;
+    });
+
+    it("others", () => {
+      sinon.stub(UserSettings, "getConfigSync").returns(
+        ok({
+          [automaticNpmInstallOption]: "others",
+        })
+      );
+      expect(getAutomaticNpmInstallSetting()).to.be.true;
+    });
+
+    it("none", () => {
+      sinon.stub(UserSettings, "getConfigSync").returns(ok({}));
+      expect(getAutomaticNpmInstallSetting()).to.be.false;
+    });
+
+    it("getConfigSync error", () => {
+      const error = returnUserError(new Error("Test"), cliSource, "Test");
+      sinon.stub(UserSettings, "getConfigSync").returns(err(error));
+      expect(getAutomaticNpmInstallSetting()).to.be.false;
+    });
+
+    it("getConfigSync exception", () => {
+      sinon.stub(UserSettings, "getConfigSync").throws("Test");
+      expect(getAutomaticNpmInstallSetting()).to.be.false;
     });
   });
 });
